Write iteration of fibonacci sequence
Wrote a function called 'fibs'. Set the first two points of the array
manually as a constant, for the sake of variables 0-2 (listed below).

Set an error that forces users to log a new number if they inputted one
less than or equal to zero. Also set a constant so that if a user inputs
one, they get 0 as a result. Inputting 2 returns the array constant listed
above.

From there, ran a for loop that starts at 2 and increments until it
reaches the number that was inputted by the user. A push is attached to
the array that adds the second-to-last number in the array with the last
number to calculate the fibonacci.