Write a recursion for the fibonacci sequence
A similar structure at first to the iteration fibonaccci script. This time
I used a ternary operator to determine if a situation was true or false.
Once we know the user did not input a number => 2, there is a spread operator
that takes the array from n-1 and adds onto it by taking the second to
last and third to last integers.